EOL of any kind removed from 'line' on input. EOL appended to every line on output, with the expception of buffers created with Mustache. Mustache created buffers depend upon the template code for EOLs - with the exception that a final EOL is added to the buffer if one is not detected. eol is now passed in as an option, unix style by default.
All test are working.

A basic form of "symCmdMacro"  is working.  The macro
allow positional argument passed in via partials.
However the arguments are all treated as strings - no data objects.

In addition to "symAddDef" which just allows strings to be added,
"symAddDefJson" and "symAddDefEval" have been added to allow data objects to be added.
The functionality is parsing as the names suggest.

Next planned for implementation:
- Macros with
  - optional array expansion without the Mustache syntax
  - optional arg conversion via Json.parse, eval, defines lookup
- Multiline commands using, e.g., /*--addDef  <name> <body over several lines>   --end*/